+++ src/main/java/fr/spring/eBoutique/project/dao/IAdresseDAO.java
@@ -5,7 +5,7 @@ import fr.spring.eBoutique.project.model.Adresse;
 
 public interface IAdresseDAO {
 
-	public Adresse getAdresse(Integer idUtilisateur);
+	public Adresse getAdresse(Long idUtilisateur);
 	public Adresse addAdresse(Adresse adresse, Integer idUtilisateur);
 	public void updateAdresse(Adresse adresse, Integer idUtilisateur);
 	public void removeAdresse(Integer idUtilisateur);
