+++ src/main/java/fr/spring/eBoutique/project/dao/IUtilisateurDAO.java
@@ -5,7 +5,7 @@ import fr.spring.eBoutique.project.model.Utilisateur;
 
 public interface IUtilisateurDAO<Login> {
 
-	public Utilisateur getUtilisateurById (Integer id);
+	public Utilisateur getUtilisateurById (Long id);
 	public Utilisateur getUtilisateurByLogin(String login);
 	public Utilisateur getUtilisateurByEmail(String email);
 	public Utilisateur  addUtilisateur(Utilisateur utilisateur);
